+++ src/components/standard/calendario.css
@@ -1,6 +1,6 @@
 .calendar-container {
   width: 600px;
-  margin: 20px auto; /* antes: 60px auto */
+  margin: 20px auto;
   background: white;
   border: 2px solid #007aff;
   border-radius: 16px;
@@ -28,7 +28,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  padding-bottom: 16px; /* sube un poco el símbolo */
+  padding-bottom: 16px; 
   color: black;
 }
 
@@ -94,4 +94,163 @@
 .done-button:hover {
   background-color: #004ea3;
 }
+
+
+.horario-modal-overlay {
+  position: fixed;
+  z-index: 1000;
+  inset: 0;
+  background: rgba(0,0,0,0.25);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.horario-modal {
+  background: #fff;
+  border-radius: 20px;
+  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.18);
+  padding: 32px 24px 24px 24px;
+  min-width: 420px;
+  min-height: 320px;
+  max-width: 95vw;
+  max-height: 95vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.horario-title {
+  text-align: center;
+  font-size: 24px;
+  margin-bottom: 6px;
+  color: #002c63;
+}
+
+.horario-date {
+  text-align: center;
+  color: #444;
+  font-size: 15px;
+  margin-bottom: 18px;
+}
+
+.horario-grid-outer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #fff;
+  border-radius: 20px;
+  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.12);
+  padding: 18px 8px;
+  margin-bottom: 24px;
+  min-width: 0;
+  min-height: 0;
+  border: 2px solid #e0e7ef;
+  max-width: 700px;
+  width: 100%;
+}
+
+.horario-grid {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.horario-row {
+  display: flex;
+  gap: 10px;
+}
+
+.horario-btn {
+  width: 90px;
+  height: 40px;
+  background: #fff;
+  border: 1px solid #bdbdbd;
+  border-radius: 8px;
+  font-size: 12px; 
+  color: #222;
+  cursor: pointer;
+  transition: background 0.2s, color 0.2s, border 0.2s;
+  outline: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-family: inherit;
+  font-weight: 640;
+}
+
+.horario-btn.selected {
+  background: #2563eb;
+  color: #fff;
+  border: 2.5px solid #2563eb;
+}
+
+.horario-btn:active {
+  filter: brightness(0.95);
+}
+
+.horario-grid-centered {
+  align-items: center;
+}
+
+.horario-row-centered {
+  justify-content: center;
+}
+
+.all-day-btn {
+  font-weight: 600;
+  letter-spacing: 0.5px;
+  background: #f0f6ff;
+  border: 2px solid #2563eb;
+}
+
+.all-day-btn.selected {
+  background: #2563eb;
+  color: #fff;
+  border: 2.5px solid #2563eb;
+}
+
+.horario-footer {
+  margin-top: 10px;
+  text-align: center;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  gap: 12px;
+}
+
+.next-btn {
+  background-color: #002c63;
+  color: white;
+  border: none;
+  padding: 8px 20px;
+  border-radius: 6px;
+  font-size: 15px;
+  cursor: pointer;
+}
+
+.next-btn:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
+
+.next-btn:hover:enabled {
+  background-color: #004080;
+}
+
+.cancel-btn {
+  background-color: transparent;
+  color: #002c63;
+  border: 1.5px solid #002c63;
+  padding: 8px 20px;
+  border-radius: 6px;
+  font-size: 15px;
+  cursor: pointer;
+  font-weight: 500;
+}
+
+.cancel-btn:hover {
+  background-color: #f0f4fa;
+}
+
