--- src/components/standard/horario.css
@@ -1,268 +1,0 @@
-.horario-container {
-  font-family: 'Segoe UI', sans-serif;
-  padding: 9px 30px; 
-  max-width: 1100px;
-  width: 100%;
-  margin: 0 auto;
-}
-
-.horario-title {
-  text-align: center;
-  font-size: 28px;
-  margin-bottom: 6px;
-  color: #002c63;
-}
-
-.horario-date {
-  text-align: center;
-  color: #444;
-  font-size: 17px;
-  margin-bottom: 30px;
-}
-
-.horario-card {
-  background-color: #ffffff;
-  border-radius: 20px;
-  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.12);
-  padding: 40px 50px;
-  display: flex;
-  gap: 70px;
-  align-items: flex-start;
-  width: 100%;
-}
-
-.inputs-container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-}
-
-.time-block {
-  background-color: #e5e5e5;
-  border-radius: 18px;
-  padding: 24px 36px;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.15);
-  border: 1px solid #ccc;
-}
-
-.time-label {
-  font-size: 16px;
-  font-weight: bold;
-  display: block;
-  margin-bottom: 12px;
-  color: #333;
-}
-
-.time-input-group {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.time-input-group input {
-  width: 200px; 
-  font-size: 30px; 
-  padding: 8px;
-  text-align: center;
-  border: none;
-  border-radius: 10px;
-  background-color: #002c63;
-  color: white;
-}
-
-.separator {
-  font-size: 32px;
-  color: #333;
-}
-
-.ampm-buttons {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.ampm-buttons button {
-  background: #f0f0f0;
-  border: 1px solid #ccc;
-  font-size: 12px;
-  padding: 4px;
-  cursor: default;
-  border-radius: 4px;
-}
-
-.time-actions {
-  margin-top: 16px;
-  display: flex;
-  justify-content: flex-end;
-  gap: 12px;
-}
-
-.cancel-btn,
-.done-btn {
-  padding: 8px 16px;
-  border-radius: 6px;
-  font-weight: 600;
-  cursor: pointer;
-  border: none;
-  font-size: 14px;
-}
-
-.cancel-btn {
-  background-color: transparent;
-  color: #002c63;
-}
-
-.done-btn {
-  background-color: #002c63;
-  color: white;
-}
-
-.schedule-visual {
-  flex: 1;
-  background-color: #fffdd7;
-  border-radius: 18px;
-  padding: 20px;
-  position: relative;
-  height: 580px;
-  display: flex;
-  align-items: flex-start;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.15);
-  border: 1px solid #ddd;
-}
-
-.hour-list {
-  display: flex;
-  flex-direction: column;
-  gap: 24px; /* Más espacio entre horas */
-  font-size: 16px; /* Más grande */
-  font-weight: 500;
-  color: #333;
-}
-
-.bar-container {
-  position: absolute;
-  left: 120px;
-  top: 20px;
-  width: 55px; /* Más ancha */
-  height: 540px;
-}
-
-.time-block-bar {
-  position: absolute;
-  width: 100%;
-  background-color: #002c63;
-  border-radius: 8px;
-}
-
-.schedule-visual::before {
-  content: "";
-  position: absolute;
-  top: 20px;
-  left: 0;
-  right: 0;
-  height: 100%;
-  background-image: repeating-linear-gradient(
-    to bottom,
-    transparent,
-    transparent 39px,
-    rgba(0, 0, 0, 0.1) 40px
-  );
-  pointer-events: none;
-  z-index: 0;
-}
-
-.horario-footer {
-  margin-top: 30px;
-  text-align: center;
-}
-
-.next-btn {
-  background-color: #002c63;
-  color: white;
-  border: none;
-  padding: 10px 24px;
-  border-radius: 6px;
-  font-size: 16px;
-  cursor: pointer;
-}
-
-.next-btn:hover {
-  background-color: #004080;
-}
-
-.horario-grid-outer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #fff;
-  border-radius: 20px;
-  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.12);
-  padding: 25px 20px;    
-  margin-bottom: 100px;
-  min-width: 0;          
-  min-height: 290px;
-  border: 2px solid #e0e7ef;
-  max-width: 1000px;       
-  width: 100%;             
-}
-
-.horario-grid {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-}
-
-.horario-row {
-  display: flex;
-  gap: 14px;
-}
-
-.horario-btn {
-  width: 110px;
-  height: 54px;
-  background: #fff;
-  border: 2px solid #bdbdbd;
-  border-radius: 10px;
-  font-size: 15px; /* reducido de 20px a 15px */
-  color: #222;
-  cursor: pointer;
-  transition: background 0.2s, color 0.2s, border 0.2s;
-  outline: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-family: inherit;
-  font-weight: 500;
-}
-
-.horario-btn.selected {
-  background: #2563eb;
-  color: #fff;
-  border: 2.5px solid #2563eb;
-}
-
-.horario-btn:active {
-  filter: brightness(0.95);
-}
-
-.horario-grid-centered {
-  align-items: center;
-}
-
-.horario-row-centered {
-  justify-content: center;
-}
-
-.all-day-btn {
-  font-weight: 600;
-  letter-spacing: 0.5px;
-  background: #f0f6ff;
-  border: 2px solid #2563eb;
-}
-
-.all-day-btn.selected {
-  background: #2563eb;
-  color: #fff;
-  border: 2.5px solid #2563eb;
-}
